Refactoring code
Lots of logic put into small succinct functions in main and modules.py split into 4 smaller module files
